# poc/contracts/backend.openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/admin/contents": {
      "post": {
        "operationId": "AdminContentController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateContentBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Draft created at version 1",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminContentView"
                }
              }
            }
          },
          "401": {
            "description": "unauthenticated",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          },
          "403": {
            "description": "forbidden; content:write is required",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          },
          "409": {
            "description": "slug_conflict for a manual slug already in use",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          },
          "413": {
            "description": "payload_too_large: request body exceeds 256 KB",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          },
          "422": {
            "description": "validation_failed with the offending field names",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a draft content",
        "tags": [
          "admin"
        ]
      },
      "get": {
        "operationId": "AdminContentController_list",
        "parameters": [
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "description": "Opaque base64url cursor."
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          },
          {
            "name": "category",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "film",
                "sorozat",
                "hir",
                "sport",
                "szorakozas",
                "egyeb"
              ]
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "draft",
                "published",
                "withdrawn"
              ]
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 200
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Editorial list without full content payloads",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminContentListView"
                }
              }
            }
          },
          "401": {
            "description": "unauthenticated",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          },
          "403": {
            "description": "forbidden; content:read is required",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          },
          "422": {
            "description": "validation_failed for invalid, unknown or repeated query fields",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          },
          "503": {
            "description": "dependency_unavailable",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List editorial content with stable cursor pagination",
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/contents/{id}": {
      "patch": {
        "operationId": "AdminContentController_patch",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchContentBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated, or the unchanged record on a no-op",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminContentView"
                }
              }
            }
          },
          "401": {
            "description": "unauthenticated",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          },
          "403": {
            "description": "forbidden; content:write is required",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          },
          "404": {
            "description": "content_not_found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          },
          "409": {
            "description": "409 version_conflict, state conflict or slug_conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          },
          "413": {
            "description": "payload_too_large: request body exceeds 256 KB",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          },
          "422": {
            "description": "validation_failed with the offending field names",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Edit a draft or withdrawn content",
        "tags": [
          "admin"
        ]
      },
      "get": {
        "operationId": "AdminContentController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Editorial view including mediaAssetId",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminContentView"
                }
              }
            }
          },
          "401": {
            "description": "unauthenticated",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          },
          "403": {
            "description": "forbidden; content:read is required",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          },
          "404": {
            "description": "content_not_found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Editorial detail in every state",
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/contents/{id}/publish": {
      "post": {
        "operationId": "AdminContentController_publish",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VersionedCommandBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Published; the DB write is committed, searchability is not implied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminContentView"
                }
              }
            }
          },
          "401": {
            "description": "unauthenticated",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          },
          "403": {
            "description": "forbidden; content:publish is required",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          },
          "404": {
            "description": "content_not_found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          },
          "409": {
            "description": "409 version_conflict, state conflict or slug_conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          },
          "413": {
            "description": "payload_too_large: request body exceeds 256 KB",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          },
          "422": {
            "description": "validation_failed with the offending field names",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Publish a draft or withdrawn content",
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/contents/{id}/withdraw": {
      "post": {
        "operationId": "AdminContentController_withdraw",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VersionedCommandBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Withdrawn; the slug stays reserved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminContentView"
                }
              }
            }
          },
          "401": {
            "description": "unauthenticated",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          },
          "403": {
            "description": "forbidden; content:publish is required",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          },
          "404": {
            "description": "content_not_found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          },
          "409": {
            "description": "409 version_conflict, state conflict or slug_conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          },
          "413": {
            "description": "payload_too_large: request body exceeds 256 KB",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          },
          "422": {
            "description": "validation_failed with the offending field names",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Withdraw a published content",
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/contents/{id}/audit": {
      "get": {
        "operationId": "AdminContentController_audit",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "description": "Opaque base64url cursor."
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 50
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Audit metadata without request bodies or field values",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContentAuditListView"
                }
              }
            }
          },
          "401": {
            "description": "unauthenticated",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          },
          "403": {
            "description": "forbidden; content:read is required",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          },
          "404": {
            "description": "content_not_found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          },
          "422": {
            "description": "validation_failed for id, cursor or query fields",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          },
          "503": {
            "description": "dependency_unavailable",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List the append-only content audit trail, newest first",
        "tags": [
          "admin"
        ]
      }
    },
    "/catalog/contents/{id}": {
      "get": {
        "operationId": "CatalogContentController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Public fields only; no actor, media asset id or audit data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicContentView"
                }
              }
            }
          },
          "404": {
            "description": "content_not_found: not published or unknown id",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          },
          "422": {
            "description": "validation_failed when the id is not a UUID",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDocument"
                }
              }
            }
          }
        },
        "summary": "Public detail of currently published content",
        "tags": [
          "catalog"
        ]
      }
    },
...
  },
  "info": {
    "title": "IndaPlay PoC backend",
    "description": "Errors are application/problem+json documents with a stable `code`. Admin routes require a verified bearer token once FEATURE_IDENTITY is on.",
    "version": "0.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "CreateContentBody": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "slug": {
            "nullable": true,
            "type": "string"
          },
          "summary": {
            "nullable": true,
            "type": "string"
          },
          "category": {
            "nullable": true,
            "type": "string"
          },
          "mediaAssetId": {
            "nullable": true,
            "type": "string"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "title"
        ],
        "additionalProperties": false
      },
...
      "AdminContentView": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          },
          "title": {
            "type": "string"
          },
          "slug": {
            "nullable": true,
            "type": "string"
          },
          "summary": {
            "nullable": true,
            "type": "string"
          },
          "category": {
            "nullable": true,
            "type": "string",
            "enum": [
              "film",
              "sorozat",
              "hir",
              "sport",
              "szorakozas",
              "egyeb"
            ]
          },
          "mediaAssetId": {
            "nullable": true,
            "type": "string"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "published",
              "withdrawn"
            ]
          },
          "version": {
            "type": "integer",
            "minimum": 0,
            "exclusiveMinimum": true,
            "maximum": 9007199254740991,
            "description": "Optimistic-concurrency token; echo it back as expectedVersion."
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$",
            "description": "ISO 8601 timestamp in UTC (Z suffix)."
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$",
            "description": "ISO 8601 timestamp in UTC (Z suffix)."
          },
          "publishedAt": {
            "nullable": true,
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$",
            "description": "ISO 8601 timestamp in UTC (Z suffix)."
          },
          "withdrawnAt": {
            "nullable": true,
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$",
            "description": "ISO 8601 timestamp in UTC (Z suffix)."
          },
          "createdBy": {
            "type": "string"
          },
          "updatedBy": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "slug",
          "summary",
          "category",
          "mediaAssetId",
          "tags",
          "status",
          "version",
          "createdAt",
          "updatedAt",
          "publishedAt",
          "withdrawnAt",
          "createdBy",
          "updatedBy"
        ],
        "additionalProperties": false
      },
...
    }
  }
}